# Conflicts:
#	DoctorAppointmentSystem/src/Controller/AddDoctor.java
#	DoctorAppointmentSystem/src/Controller/DashBoard.java
#	DoctorAppointmentSystem/src/View/MainDashboard.fxml
#	src/com/asd/framework/DatabaseConnection/Db/Test.java

+++ src/com/asd/framework/Reminder/ReminderSMSStrategy.java
@@ -13,7 +13,7 @@ public class ReminderSMSStrategy extends ReminderDefaultStrategy {
 	}
 
 	@Override
-	public boolean send(Integer receiverId, String text) {
+	public boolean send(Long receiverId, String text) {
 		//Connect to SMS Gateway
 		return true;
 	}
